mrowki znowu

+++ FR_09_12/Program.cs
@@ -1,31 +1,25 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FR_09_12
 {
     class Program
     {
-
         static void Main(string[] args)
         {
-
             int Q = int.Parse(Console.ReadLine());
 
-            int[,] tab = new int[2,Q];
+            long[,] tab = new long[2,Q];
 
             for (int i = 0; i < Q; i++)
             {
-                int[] x = GetNumber1();
+                long[] x = GetNumber1();
                 tab[0, i] = x[0];
                 tab[1, i] = x[1];
             }
 
-            int currentMin = tab[1, Q - 1];
+            long currentMin = tab[1, Q - 1];
             bool isAdjacent = true;
-            int cnt = 1;
+            long cnt = 1;
 
             for (int i = Q-2; i >=0; i--)
             {
@@ -51,10 +45,10 @@ namespace FR_09_12
             Console.WriteLine(cnt);
         }
 
-        static int[] GetNumber1()
+        static long[] GetNumber1()
         {
             string[] tab = (Console.ReadLine()).Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
-            return Array.ConvertAll(tab,int.Parse);
+            return Array.ConvertAll(tab,long.Parse);
         }
     }
 }
